Fix make test issue and
The error before "Transform function not recognized" was due to leaving the functions from the ingest template in, which were just examples, but they had a different tag like "ingest_by_record" so koza thought that was a tag it was supposed to run, and didn't see a transform function for it.

Make test log:
============ 251 passed, 14 warnings in 98.16s (0:01:38) =============
uv run codespell --skip="./data/*,**/site-packages" --ignore-words=.codespellignore
uv run ruff check
All checks passed!